........
  r83751 | mark.dickinson | 2010-08-06 10:36:57 +0100 (Fri, 06 Aug 2010) | 1 line

  Issue #9526:  Remove outdated casts to int that were preventing the array module from working correctly with arrays > 2GB.
........
  r83752 | mark.dickinson | 2010-08-06 10:38:58 +0100 (Fri, 06 Aug 2010) | 1 line

  Misc/NEWS entry for r83751.
........

+++ Modules/arraymodule.c
@@ -798,7 +798,7 @@ array_iter_extend(arrayobject *self, PyObject *bb)
         return -1;
 
     while ((v = PyIter_Next(it)) != NULL) {
-        if (ins1(self, (int) Py_SIZE(self), v) != 0) {
+        if (ins1(self, Py_SIZE(self), v) != 0) {
             Py_DECREF(v);
             Py_DECREF(it);
             return -1;
@@ -1090,7 +1090,7 @@ the buffer length in bytes.");
 static PyObject *
 array_append(arrayobject *self, PyObject *v)
 {
-    return ins(self, (int) Py_SIZE(self), v);
+    return ins(self, Py_SIZE(self), v);
 }
 
 PyDoc_STRVAR(append_doc,
